+++ podcasts/ProfileViewController.swift
@@ -343,19 +343,23 @@ class ProfileViewController: PCViewController, UITableViewDataSource, UITableVie
 
     @objc private func whatsNewDismissed() {
         showGeneralSettingsIfNeeded()
+        showHeadphoneControlsFromWhatsNew()
     }
 
     private func showGeneralSettingsIfNeeded() {
-        if AnnouncementFlow.shared.isShowingAutoplayOption {
+        if AnnouncementFlow.current == .autoPlay {
             let generalSettingsViewController = GeneralSettingsViewController()
             navigationController?.pushViewController(generalSettingsViewController, animated: true)
         }
+    }
 
-        if AnnouncementFlow.shared.isShowingBookmarksOption {
-            let controller = HeadphoneSettingsViewController()
-            navigationController?.pushViewController(controller, animated: true)
-            AnnouncementFlow.shared.isShowingBookmarksOption = false
-        }
+    // Pushes to the headphone controls if shown from the what's new
+    private func showHeadphoneControlsFromWhatsNew() {
+        guard AnnouncementFlow.current == .bookmarksProfile else { return }
+
+        let controller = HeadphoneSettingsViewController()
+        navigationController?.pushViewController(controller, animated: true)
+        AnnouncementFlow.current = .none
     }
 }